requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-storage
setuptools
gunicorn
psycopg2-binary
python-jose
passlib
opencv-python-headless
tensorflow
numpy